requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
python-dotenv
httpx
pytesseract
PyMuPDF
Pillow
